Refactor pattern provider
