+++ src/_providers/index.ts
@@ -1,3 +1,2 @@
-export * from "./patterns";
 export * from "./postgrest";
 export * from "./auth";
