+++ src/pages/dev/concepts/editor/+Page.ts
@@ -1,5 +1,5 @@
 import SectionEditor from "@macrostrat-web/section-editor-demo";
-import { PatternProvider } from "~/_providers";
+import { PatternProvider } from "~/_providers.client";
 import h from "@macrostrat/hyper";
 
 export function Page() {
