+++ src/pages/dev/concepts/multiscale-figure/+Page.ts
@@ -1,6 +1,6 @@
 import h, { C, compose } from "@macrostrat/hyper";
 import { APIProvider, useAPIResult } from "@macrostrat/ui-components";
-import { PatternProvider } from "~/_providers";
+import { PatternProvider } from "~/_providers.client";
 import {
   AlignedMeasurementProvider,
   FilteredMeasurementProvider,
